Add warehouse logic

+++ src/main/java/com/supermarket/supermarket/dto/ProductRequest.java
@@ -19,7 +19,7 @@ public class ProductRequest {
     private BigDecimal price;
 
     @NotNull
-    private int expirationDate;
+    private BigDecimal expirationDate;
 
     @NotNull
     private Long category;
